2025OSSPr/requirements.txt
altair==5.5.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
attrs==25.3.0
blinker==1.9.0
branca==0.8.0
cachetools==5.5.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
display==1.0.0
et_xmlfile==2.0.0
executing==2.1.0
folium==0.18.0
fonttools==4.54.1
geokakao==0.1.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
googletrans==4.0.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jpype1==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.7
knnimpute==0.1.0
konlpy==0.6.0
koreanize-matplotlib==0.1.1
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mlxtend==0.23.3
narwhals==1.37.1
nest-asyncio==1.6.0
numpy==2.1.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.2
plotly==5.24.1
prompt_toolkit==3.0.47
protobuf==6.30.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydeck==0.9.1
Pygments==2.18.0
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
statsmodels==0.14.4
streamlit==1.45.0
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.2.3
wcwidth==0.2.13
wordcloud==1.9.4
xyzservices==2024.9.0
